+++ src/models/service_route_model.rs
@@ -1,7 +1,7 @@
 pub struct ServiceRoute {
-    pub id: usize,
-    pub image_fk: usize,      // a reference to the image,
+    pub id: i32,
+    pub image_fk: i32,        // a reference to the image,
     pub prefix: String,       //prefix of the service
     pub exposed_port: String, // port where the service within the container would listen,
-    pub segments: usize,
+    pub segments: i32,
 }
